backend/requirements.txt
fastapi
uvicorn
anthropic
pocketbase
python-dotenv
pytest
pytest-asyncio
httpx
pydantic
